ref(events-v2) Split up the EventDetails view better
Split the logic up so that there is only one stateful component and
several controlled components that focus on presentation only.

Refs SEN-648
